+++ app/src/test/java/eu/kanade/tachiyomi/data/backup/MangaRestoreCategoriesTest.kt
@@ -9,6 +9,7 @@ import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Test
+import reikai.domain.merge.RestoreMergeGroups
 import tachiyomi.data.Database
 import tachiyomi.domain.category.interactor.GetCategories
 import tachiyomi.domain.category.model.Category
@@ -69,7 +70,7 @@ class MangaRestoreCategoriesTest {
             getTracks = mockk(relaxed = true),
             insertTrack = mockk(relaxed = true),
             fetchInterval = mockk(relaxed = true),
-            mergeGroupRepository = mockk(relaxed = true),
+            restoreMergeGroups = RestoreMergeGroups(mockk(relaxed = true)),
             mangaMetadataRepository = mockk(relaxed = true),
             setCustomMangaInfo = mockk(relaxed = true),
         )